requirements.txt

flask
flask-cors
ultralytics
pymupdf
PyPDF2
Pillow
